pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trucking-document-management"
version = "1.0.0"
description = "Flask-based trucking document management system with Telegram integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator==2.2.0",
    "flask==3.1.0",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.1",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.9",
    "python-telegram-bot==21.8",
    "sendgrid==6.11.0",
    "werkzeug==3.1.3",
    "wtforms==3.2.1"
]

[tool.setuptools]
py-modules = ["main"]